+++ e2e/e2etests/e2etests.go
@@ -147,7 +147,10 @@ const (
 	/*
 	 Stateful precompiled contracts tests
 	*/
-	TestZetaPrecompilesPrototypeName = "precompile_contracts_prototype"
+	TestPrecompilesPrototypeName                = "precompile_contracts_prototype"
+	TestPrecompilesPrototypeThroughContractName = "precompile_contracts_prototype_through_contract"
+	TestPrecompilesStakingName                  = "precompile_contracts_staking"
+	TestPrecompilesStakingThroughContractName   = "precompile_contracts_staking_through_contract"
 )
 
 // AllE2ETests is an ordered list of all e2e tests
@@ -814,9 +817,27 @@ var AllE2ETests = []runner.E2ETest{
 	 Stateful precompiled contracts tests
 	*/
 	runner.NewE2ETest(
-		TestZetaPrecompilesPrototypeName,
+		TestPrecompilesPrototypeName,
 		"test stateful precompiled contracts prototype",
 		[]runner.ArgDefinition{},
-		TestPrecompilesRegular,
+		TestPrecompilesPrototype,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesPrototypeThroughContractName,
+		"test stateful precompiled contracts prototype through contract",
+		[]runner.ArgDefinition{},
+		TestPrecompilesPrototypeThroughContract,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesStakingName,
+		"test stateful precompiled contracts staking",
+		[]runner.ArgDefinition{},
+		TestPrecompilesStaking,
+	),
+	runner.NewE2ETest(
+		TestPrecompilesStakingThroughContractName,
+		"test stateful precompiled contracts staking through contract",
+		[]runner.ArgDefinition{},
+		TestPrecompilesStakingThroughContract,
 	),
 }